renames ComponentWithLabel to just Label and flattens file structure
