+++ src/layout/Checkboxes/CheckboxesContainerComponent.tsx
@@ -4,7 +4,7 @@ import { Checkbox } from '@digdir/designsystemet-react';
 import cn from 'classnames';
 
 import { AltinnSpinner } from 'src/components/AltinnSpinner';
-import { LabelContent } from 'src/features/label/LabelContent/LabelContent';
+import { LabelContent } from 'src/features/label/LabelContent';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
 import { useGetOptions } from 'src/features/options/useGetOptions';
