+++ src/layout/Dropdown/DropdownComponent.tsx
@@ -6,7 +6,7 @@ import { AltinnSpinner } from 'src/components/AltinnSpinner';
 import { ConditionalWrapper } from 'src/components/ConditionalWrapper';
 import { DeleteWarningPopover } from 'src/components/molecules/DeleteWarningPopover';
 import { FD } from 'src/features/formData/FormDataWrite';
-import { ComponentWithLabel } from 'src/features/label/ComponentWithLabel/ComponentWithLabel';
+import { Label } from 'src/features/label/Label';
 import { Lang } from 'src/features/language/Lang';
 import { useLanguage } from 'src/features/language/useLanguage';
 import { useGetOptions } from 'src/features/options/useGetOptions';
@@ -54,12 +54,12 @@ export function DropdownComponent({ node, isValid, overrideDisplay }: IDropdownP
 
   const withLabel = (Component) =>
     overrideDisplay?.renderedInTable !== true ? (
-      <ComponentWithLabel
+      <Label
         {...node.item}
         renderLabelAs='label'
       >
         {Component}
-      </ComponentWithLabel>
+      </Label>
     ) : (
       Component
     );
